requirements.txt
# Environment
pettingzoo>=1.24.0
gymnasium>=0.29.0
pygame>=2.5.0

# RL
torch>=2.0.0
stable-baselines3>=2.0.0
tensorboard>=2.13.0

# Logging
wandb>=0.15.0
pandas>=2.0.0

# Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
scipy>=1.11.0

# Utils
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
